requirements.txt
fastapi
numpy
python-dotenv
